Move trotter to grid_ham/trotter.
