+++ qualtran/bloqs/chemistry/trotter/grid_ham/__init__.py
