+++ qualtran/bloqs/chemistry/trotter/grid_ham/kinetic.py
@@ -28,7 +28,7 @@ from qualtran import (
     SoquetT,
 )
 from qualtran.bloqs.arithmetic import SumOfSquares
-from qualtran.bloqs.chemistry.trotter.qvr import QuantumVariableRotation
+from qualtran.bloqs.chemistry.trotter.grid_ham.qvr import QuantumVariableRotation
 
 
 @frozen
@@ -85,6 +85,6 @@ def _kinetic_energy() -> KineticEnergy:
 
 _KINETIC_ENERGY = BloqDocSpec(
     bloq_cls=KineticEnergy,
-    import_line='from qualtran.bloqs.chemistry.trotter.kinetic import KineticEnergy',
+    import_line='from qualtran.bloqs.chemistry.trotter.grid_ham.kinetic import KineticEnergy',
     examples=(_kinetic_energy,),
 )
