+++ qualtran/bloqs/chemistry/trotter/grid_ham/qvr_test.py
@@ -11,7 +11,7 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-from qualtran.bloqs.chemistry.trotter.qvr import _qvr
+from qualtran.bloqs.chemistry.trotter.grid_ham.qvr import _qvr
 
 
 def test_kinetic_energy(bloq_autotester):
